User commit

+++ src/main/java/com/example/demo/entity/Student.java
@@ -13,15 +13,15 @@ import jakarta.persistence.Id;
 public class Student {
    @Id
    @GeneratedValue(strategy=GenerationType.IDENTITY)
-   private int id;
+   private Long id;
    private String name;
    private String email;
    private LocalDate dob;
    private float cgpa;
-   public int getId() {
+   public Long getId() {
     return id;
    }
-   public void setId(int id) {
+   public void setId(Long id) {
     this.id = id;
    }
    public String getName() {
@@ -48,7 +48,7 @@ public class Student {
    public void setCgpa(float cgpa) {
     this.cgpa = cgpa;
    }
-   public Student(int id, String name, String email, LocalDate dob, float cgpa) {
+   public Student(Long id, String name, String email, LocalDate dob, float cgpa) {
     this.id = id;
     this.name = name;
     this.email = email;
